add toc to projects page
the template invoking the toc code expects a div of class "entry-content",
which is not created for individual pages, so we manually wrap this page
in such a div. consider changing the template to add the div instead.

maruku didn't seem to like a <http://foo> style link inside the div,
so I switched it to [this style](http://foo).

Deleted old content while I was at it.